Initial overhaul
Changed the layout of the page itself so that the correct parts of the
page scroll. Removed the large header. Added an additional layer for
extra graphics. Added a proper tilesheet (open source). Added hitbox
edit mode, for adjusting whether or not a tile is hitable. Changed build
format: For each tile, the spritesheet index and hitbox status of the
tile is given.